requirements.txt
streamlit
scikit-learn
numpy
Pillow
joblib